bug delete image

+++ src/modules/categoria/DeleteCategoria/DeleteCategoriaUseCase.js
@@ -16,18 +16,21 @@ class DeleteCategoriaUseCase {
       throw new AppError("Sem permissão.")
     }
 
+    
     const categoriaAntiga = await prisma.categoria.findFirst({
       where: {
         id,
       },
     });
-
+    
     if (!categoriaAntiga) {
       throw new AppError("Esta categoria não existente.");
     }
-
-    await diskStorage.deleteFile(categoriaAntiga.imagem.split('.app/')[1])
-
+    
+    if (categoriaAntiga.imagem.includes('https://cardapiobackend.up.railway.app/')) {
+      await diskStorage.deleteFile(categoriaAntiga.imagem.split('.app/')[1])
+    }
+    
     await prisma.categoria.delete({
       where: {
         id,
